requirements.txt
numpy
scipy
scikit-learn
umap-learn
matplotlib
seaborn
joblib
pandas
